do not use IAIMP2Controller::AIMP_PLS_Info, it is was deprecated in AIMP2, and in AIMP3 it returns incorrect info before user manualy activate playlist in player's main window.

+++ src/aimp/playlist.cpp
@@ -24,7 +24,7 @@ Playlist::Playlist()
 
 Playlist::Playlist( const CHAR* title,
                     DWORD file_count,
-                    DWORD duration,
+                    INT64 duration,
                     INT64 size_of_all_entries_in_bytes,
                     PlaylistID id
                    )
